fix: rename rest/typing.py to avoid shadowing Python stdlib
