+++ rest/agent/tools/github_issue.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from rest.agent.tools.tool import Tool
 from rest.tools.github import GitHubClient
-from rest.typing import ActionType
+from rest.rest_types import ActionType
 
 
 class CreateGitIssue(Tool):
