+++ rest/driver/telemetry_logic.py
@@ -24,7 +24,7 @@ from rest.config import (
 )
 from rest.dao.sqlite_dao import TraceRootSQLiteClient
 from rest.service.provider import ObservabilityProvider
-from rest.typing import Operation
+from rest.rest_types import Operation
 from rest.utils.pagination import PaginationHelper
 from rest.utils.provider import get_observe_provider
 from rest.utils.trace_cache import TraceCacheHelper
